+++ k256/Cargo.toml
@@ -26,7 +26,6 @@ ecdsa-core = { version = "=0.15.0-pre", package = "ecdsa", optional = true, defa
 hex-literal = { version = "0.3", optional = true }
 serdect = { version = "0.1", optional = true, default-features = false }
 sha2 = { version = "0.10", optional = true, default-features = false }
-sha3 = { version = "0.10", optional = true, default-features = false }
 
 [dev-dependencies]
 blobby = "0.3"
@@ -37,6 +36,7 @@ num-bigint = "0.4"
 num-traits = "0.2"
 proptest = "1.0"
 rand_core = { version = "0.6", features = ["getrandom"] }
+sha3 = { version = "0.10", default-features = false }
 
 [features]
 default = ["arithmetic", "ecdsa", "pkcs8", "schnorr", "std"]
@@ -51,7 +51,6 @@ ecdsa = ["arithmetic", "ecdsa-core/sign", "ecdsa-core/verify", "sha256"]
 expose-field = ["arithmetic"]
 hash2curve = ["arithmetic", "elliptic-curve/hash2curve"]
 jwk = ["elliptic-curve/jwk"]
-keccak256 = ["digest", "sha3"]
 pem = ["ecdsa-core/pem", "elliptic-curve/pem", "pkcs8"]
 pkcs8 = ["ecdsa-core/pkcs8", "elliptic-curve/pkcs8"]
 schnorr = ["arithmetic", "sha256"]
@@ -60,7 +59,7 @@ sha256 = ["digest", "sha2"]
 test-vectors = ["hex-literal"]
 
 [package.metadata.docs.rs]
-features = ["ecdh", "ecdsa", "keccak256", "schnorr"]
+features = ["ecdh", "ecdsa", "schnorr"]
 rustdoc-args = ["--cfg", "docsrs"]
 
 [[bench]]
